A proposal for entering the Voltaire phase
We're heading into the age of Voltaire, laying down the foundations for
decentralized decision-making. This CIP describes a mechanism for
on-chain governance that will underpin the Voltaire phase of Cardano.
The document builds on and extends the original Cardano governance
scheme that was based on a fixed number of governance keys.
It aims to provide a first step that is both valuable and technically
achievable in the near term as part of the proposed Voltaire governance
system.

It also seeks to act as a jumping-off point for continuing community
input, including on appropriate threshold settings and other on-chain
settings. Subsequent proposals may adapt and extend this proposal to
meet emerging governance needs.